+++ app/components/itemFlatList/index.tsx
@@ -1,7 +1,9 @@
 import React from "react"
 import { Image, Text, View } from "react-native";
+import Material from 'react-native-vector-icons/MaterialIcons';
 
 import useStyle from "./styles";
+import { useTheme } from "@react-navigation/native";
 
 interface ItemFlatListProps {
   title: string;
@@ -20,10 +22,13 @@ const ItemFlatList: React.FC<ItemFlatListProps> = ({
   subtitle, 
 }) => {
   const styles = useStyle();
+  const { colors } = useTheme();
+
   return (
     <View style={styles.mainContainer}>
     <View style={styles.iconView}>
-      <Image style={styles.iconImage} source={icon ? {uri: icon} : require('../../../public/not_found.png')} />
+      <Material name={icon || 'error'} size={30} color={colors.text} />
+      {/* <Image style={styles.iconImage} source={icon ? {uri: icon} : require('../../../public/not_found.png')} /> */}
     </View>
     <View style={styles.titleView}>
       <Text style={styles.titleText}>{title}</Text>