+++ app/models/interfaces/Category.ts
@@ -4,5 +4,6 @@ export default interface Category {
   _id?: Realm.BSON.UUID;
   name: string;
   iconId: number;
+  iconName?: string;
   budget: number;
 }