+++ app/models/schemas/CategorySchema.ts
@@ -8,13 +8,15 @@ export default class Category extends Realm.Object<Category> implements ICategor
     public _id: Realm.BSON.UUID,
     public name: string, 
     public iconId: number, 
+    public iconName: string, 
     public budget: number
   ){
-    super(realm, { _id, name, iconId, budget });
+    super(realm, { _id, name, iconId, iconName, budget });
 
     this._id = _id;
     this.name = name;
     this.iconId = iconId;
+    this.iconName = iconName;
     this.budget = budget;
   }
 
@@ -24,6 +26,7 @@ export default class Category extends Realm.Object<Category> implements ICategor
       _id: { type: 'uuid', default: () => new Realm.BSON.UUID() },
       name: 'string',
       iconId: 'int',
+      iconName: 'string?',
       budget: 'double'
     },
     primaryKey: '_id'