+++ app/screens/balance/index.tsx
@@ -9,6 +9,7 @@ import MainMenu from "./components/mainMenu";
 import { BalancePanel, ItemFlatList, SectionHeader } from "../../components";
 
 import { listImgBase64 } from "../../resources/static/categoriesImages";
+import { DEFAULT_CATEGORY_ICONS } from '../../resources/values/consts';
 import { DEFAULT_BLACK, DEFAULT_BUTTON_COLOR, DEFAULT_RED } from "../../resources/values/colors";
 
 import { styles } from "./styles";
@@ -30,14 +31,13 @@ const BalanceScreen: React.FC<BalanceScreenProps> = ({ navigation }) => {
   } = useBalance(navigation);
 
   const renderCategories = (balance: Balance) => {
-    console.log(balance.category.iconId)
     return (
       <TouchableOpacity key={balance._id!.toString()} onPress={() => navigation.navigate('DetailCategory', { balance })}>
         <ItemFlatList           
           title={balance.category.name} 
           value={convertToMoney(balance.category.budget - balance.totalExpenses)} 
           valueColor={(balance.category.budget - balance.totalExpenses) < 0 ? DEFAULT_RED : DEFAULT_BLACK}
-          icon={listImgBase64.find((imgBase64) => imgBase64.id === balance.category.iconId)?.data}
+          icon={DEFAULT_CATEGORY_ICONS.find((image) => image.id === balance.category.iconId)?.name}
           subtitle={`Budget: ${convertToMoney(balance.category.budget)} \nExpenses: ${convertToMoney(balance.totalExpenses)} `} 
         />
       </TouchableOpacity>