+++ app/screens/detailCategory/components/detailedTab/index.tsx
@@ -9,6 +9,7 @@ import { ItemFlatList, SectionHeader } from "../../../../components";
 
 import Transaction from "../../../../models/schemas/TransactionSchema";
 import RootStackParamList from "../../../../models/interfaces/RootScreensParams";
+import { DEFAULT_CATEGORY_ICONS } from "../../../../resources/values/consts";
 
 interface DetailedTabComponentProps {
   navigation: NativeStackNavigationProp<RootStackParamList, 'DetailCategory' | 'Transactions'>;
@@ -26,7 +27,7 @@ const DetailedTabComponent: React.FC<DetailedTabComponentProps> = ({ navigation,
       <ItemFlatList 
         title={transaction.name} 
         value={convertToMoney(transaction.value)} 
-        icon={listImgBase64.find((imgBase64) => imgBase64.id === transaction.balance.category.iconId)?.data}
+        icon={DEFAULT_CATEGORY_ICONS.find((image) => image.id === transaction.balance.category.iconId)?.name}
         subtitle={transaction.createdAt} 
       />
     </TouchableOpacity>