+++ app/screens/transactions/index.tsx
@@ -11,7 +11,7 @@ import { convertToMoney } from "../../utils/string.util";
 import { BottomButton, ItemFlatList, SearchInput, SectionHeader } from "../../components";
 
 import { listImgBase64 } from "../../resources/static/categoriesImages";
-import { SEARCH_TEXT_MINIMUM_LENGTH } from "../../resources/values/consts";
+import { DEFAULT_CATEGORY_ICONS, SEARCH_TEXT_MINIMUM_LENGTH } from "../../resources/values/consts";
 
 type TransactionsScreenProps = NativeStackScreenProps<RootStackParamList, 'Transactions'>;
 
@@ -49,7 +49,7 @@ const TransactionsScreen: React.FC<TransactionsScreenProps> = ({ navigation }) =
       <ItemFlatList 
         title={transaction.name} 
         value={convertToMoney(transaction.value)} 
-        icon={listImgBase64.find((imgBase64) => imgBase64.id === transaction.balance.category.iconId)?.data}
+        icon={DEFAULT_CATEGORY_ICONS.find((image) => image.id === transaction.balance.category.iconId)?.name}
         subtitle={transaction.createdAt} 
       />
     </TouchableOpacity>